Split out a library crate, add routing, add error handling
All of this appears to be a common pattern for organizing command line
programs in rust. Multiple tutorials suggest this organization, as does
the rust book itself.

I don't love the repetition of the 'magic' strings for the subcommands
and args, but can worry about a way to DRY that up later if necessary.