Derive FrontMatter::default, define FrontMatter::new
Now that FrontMatter.created is an Option (the prior commit did this) we
can just derive default. This changes the default implentation slightly,
as default created is None, rather than Utc::now().

Prior implementation of default (with created == Utc::now()) is renamed
to new. This better matches expected functionality as we would want to
set created to now when a note is created.