Naive first attempt at implementing link subcommand
This is a reasonable start but it still has a few issues.

First, I don't think it's perfectly atomic in how it writes the linked
files. This may mean that there's the possibility of data loss if the
file opens and truncates successfully. But then the subsequent file
write fails, leaving just an empty file. A better approach is probably
writing a temp file, and then copying that over the file that is getting
modified.

Second, linking repeatedly is going to repeatedly list the same paths in
the links_in and links_out front matter fields.

Third, this just needs more tests. There are just a lot of ways to screw
up file writing and existing coverage doesn't guard this enough.