Simplistic first attempt at an edit flag for new subcommand
I'm not totally certain this is the right way to be spawning a child
process and handling what it can do. The error messaging could be better
for the case when there's no env variable defined. And I'm not sure how
I feel about using an env variable, as opposed to an argument or a
config file or some combination of all three.

But it's a start...