Change to an undirected linking system
The is a breaking change. It replaces the front matter fields
"links_in" and "links_out" with a simpler "links" field. The idea is
that drawing a knowledge graph of a notebook doesn't really require a
directed graph and the simpler front matter and lower complexity of
undirected links is therefore desirable.