Make the graph command consider links in note markdown
This extends the current behavior of graph, which already looks at the
front-matter of notes (specifically, the links field) to build the
graph.

Note that while the current implementation works (I think :D), it is
somewhat naive, as it pulls all note content into memory simultaneously
in order to do this.