Implement the backlink command
There was a trade-off to be made here between memory usage and number of
file writes. I opted to incur more file writes for less total memory
usage since I think that's the trade-off that is desirable in situations
with really large note collections, which are the situations where this
could matter.