Modify mv subcommand to use the ignore crate
This makes it so that when mv looks at other notes for links that need
to be updated, it follows the same behavior as tags currently has by
ignoring hidden files, files in .ignore, and .gitignore.

This also means that notes in subdirectories are searched for links to
the moved file, but the behavior there is still poorly defined,
specifically because links don't consider about relative paths right
now.