Change use of ignore crate
This change moves responsibility for filtering for markdown files out of
the ignore crate matchers. The filtering is instead done later over the
directory entries that ignore returns.

This is probably marginally less performant as ignore's threads will
pass entries that they previously would have skipped and thos entries
will then be filtered serially.

But, it's also much simpler to implement and easier to grok. And if in
the future there are flags to modify ignore's args, it will be easier to
incorporate those into this code.