Modify error handling for backlink subcommand
Similar to what was just done for tags, this will make backlink log
errors that occur as it reads/writes from files. These errors will
not stop on-going execution, since it still might be possible to
backlink other files.

If an error is encountered, backlink uses exit code 1 since it's
essentially a partial failure.