Update dependencies, check in Cargo.lock, fix a bug
Somewhere along the way, serde or serde-yaml updated so that they now
add a newline to the end of serialized front matter. FrontMatter was
compensating for the prior lack of this behavior, so that's no longer
necessary and has been removed.

Also, I think it makes sense to have Cargo.lock checked into git here
because zeke is a executable CLI rather than a library.