Upgrade serde-yaml to latest version
This broke a lot of tests because serde-yaml switched from the yaml-rust
crate to a libyaml based backend and there are minor differences in how
these different backends format yaml.

In particular, the older version was quoting timestamps (e.g. in
created) and was indenting list entries (e.g. in tags and links) whereas
the newer version does not quote timestamps or indent list entries.